refactor(spaces): extract shared inspect dispatch + reject empty slug
Address PR #251 review:

- Extract `parse_view_range`, `InspectTool`, and inspect dispatch into
  a new `core/src/toolset/space_inspect.rs` (`pub(crate)`); admin and
  top-level spaces tools both call `dispatch_inspect(...)`. Removes
  the byte-for-byte duplication flagged by Bugbot.
- Add `require_space_op` helper that converts `SpaceFs::{write,delete,
  move}_file` `Ok(None)` (only reachable for an empty slug) into an
  `InvalidArgument` error, so `write/delete/move` can no longer
  silently report success on a malformed request.
